topic/chatbot/table-augmented-generation/requirements.txt
python-dotenv
openai>=1.0.0
crate
pandas
ipywidgets
